% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsCellNet.R
\docType{data}
\name{dbloc}
\alias{dbloc}
\title{tsCellNet}
\format{
An object of class \code{character} of length 1.
}
\usage{
dbloc
}
\arguments{
\item{datloc}{three input file path}

\item{expnam}{expression file name}

\item{timnam}{time-point file name}

\item{typnam}{cell-type file name}

\item{resloc}{result store path}

\item{mainlab}{result prefix name}

\item{species}{species name}

\item{controltime}{the starting point name}

\item{treattime}{if or not need to compare cell-cell interactions with starting point name. If not set the value as "total".}
}
\description{
construction time-series cell-cell interactions using scRNA-seq data
}
\examples{
out<-datainput(datloc,expnam,typnam,timnam,"FALSE",resloc)
inpexp<-as.data.frame(out$exp) #[,kep]
inptyp<-as.data.frame(out$typ)
colnames(inptyp)<-colnames(out$typ)
rownames(inptyp)<-rownames(out$typ) 
inptim<-as.data.frame(out$tim) 
usedcolo<-definecolo(tsccnetcol,inptyp$Type,resloc,mainlab)
sankey(inptim$Time,inptyp$Type,usedcolo,mainlab,resloc)
memsercpros<-preparepairs(inpexp,species,dbloc,mainlab)
singlenet(inpexp,inptyp,"Type",inptim,"Time",memsercpros,usedcolo,"FALSE",dbloc,mainlab) 
developcorrelation(inptim,usedcolo,resloc)
comparednet(controltime,treattime,usedcolo,resloc)
importance_plot(resloc,"row","outgoing","None")
}
\keyword{datasets}
